@@ -22,12 +22,15 @@ export default class ButtonContainer extends React.PureComponent {
   };
 
   componentDidMount() {
-    this._getMainColor(this.props.current);
+    const { current } = this.props;
+    this._getMainColor(current);
   }
 
   componentDidUpdate(prevProps) {
-    if (prevProps.current !== this.props.current) {
-      this._getMainColor(this.props.current);
+    const { current } = this.props;
+
+    if (prevProps.current !== current) {
+      this._getMainColor(current);
     }
   }
 
@@ -100,7 +103,7 @@ export default class ButtonContainer extends React.PureComponent {
 
   renderPrevBtn = () => {
     const { prefixCls, prevBtn } = this.props;
-
+    const { prevBtnColor } = this.state;
     return (
       <div className={`${prefixCls}-button-wrap`}>
         {prevBtn ? (
@@ -119,7 +122,7 @@ export default class ButtonContainer extends React.PureComponent {
             className={`${prefixCls}-button`}
           >
             <path
-              fill={this.state.prevBtnColor}
+              fill={prevBtnColor}
               d="M31.7 239l136-136c9.4-9.4 24.6-9.4 33.9 0l22.6 22.6c9.4 9.4 9.4 24.6 0 33.9L127.9 256l96.4 96.4c9.4 9.4 9.4 24.6 0 33.9L201.7 409c-9.4 9.4-24.6 9.4-33.9 0l-136-136c-9.5-9.4-9.5-24.6-.1-34z"
             />
           </svg>
@@ -130,6 +133,8 @@ export default class ButtonContainer extends React.PureComponent {
 
   renderNextBtn = () => {
     const { prefixCls, nextBtn } = this.props;
+    const { nextBtnColor } = this.state;
+
     return (
       <div className={`${prefixCls}-button-wrap`}>
         {nextBtn ? (
@@ -148,7 +153,7 @@ export default class ButtonContainer extends React.PureComponent {
             className={`${prefixCls}-button`}
           >
             <path
-              fill={this.state.nextBtnColor}
+              fill={nextBtnColor}
               d="M224.3 273l-136 136c-9.4 9.4-24.6 9.4-33.9 0l-22.6-22.6c-9.4-9.4-9.4-24.6 0-33.9l96.4-96.4-96.4-96.4c-9.4-9.4-9.4-24.6 0-33.9L54.3 103c9.4-9.4 24.6-9.4 33.9 0l136 136c9.5 9.4 9.5 24.6.1 34z"
             />
           </svg>
